requirements.txt
ipykernel
numpy
pandas
flask
matplotlib
seaborn
memory_profiler
streamlit
scikit-learn
openpyxl
Pillow
openai
python-dotenv
transformers
Torch
datetime